Rename the middleware classes to match the naming used in the SDK documentation
