+++ lib/Raven/Middleware/ExceptionInterfaceMiddleware.php
@@ -17,12 +17,11 @@ use Raven\Event;
 use Raven\Stacktrace;
 
 /**
- * This middleware collects information about the error or exception that
- * generated the event.
+ * This middleware collects information about the thrown exceptions.
  *
  * @author Stefano Arlandini <sarlandini@alice.it>
  */
-final class StacktraceDataCollectorMiddleware
+final class ExceptionInterfaceMiddleware
 {
     /**
      * @var Client The Raven client
@@ -52,10 +51,35 @@ final class StacktraceDataCollectorMiddleware
      */
     public function __invoke(Event $event, callable $next, ServerRequestInterface $request = null, \Exception $exception = null, array $payload = [])
     {
-        $config = $this->client->getConfig();
+        // Do not override the level if it was set explicitly by the user
+        if (!isset($payload['level']) && $exception instanceof \ErrorException) {
+            $event = $event->withLevel($this->client->translateSeverity($exception->getSeverity()));
+        }
 
-        if (null !== $exception && $config->getAutoLogStacks() && !$config->isExcludedException($exception)) {
-            $event = $event->withStacktrace(Stacktrace::createFromBacktrace($this->client, $exception->getTrace(), $exception->getFile(), $exception->getLine()));
+        if (null !== $exception) {
+            $exceptions = [];
+            $currentException = $exception;
+
+            do {
+                if ($this->client->getConfig()->isExcludedException($currentException)) {
+                    continue;
+                }
+
+                $data = [
+                    'type' => get_class($currentException),
+                    'value' => $this->client->getSerializer()->serialize($currentException->getMessage()),
+                ];
+
+                if ($this->client->getConfig()->getAutoLogStacks()) {
+                    $data['stacktrace'] = Stacktrace::createFromBacktrace($this->client, $currentException->getTrace(), $currentException->getFile(), $currentException->getLine());
+                }
+
+                $exceptions[] = $data;
+            } while ($currentException = $currentException->getPrevious());
+
+            $exceptions = array_reverse($exceptions);
+
+            $event = $event->withException($exceptions);
         }
 
         return $next($event, $request, $exception, $payload);
